Switch to use headerbar
This patch tries to refresh the UI to the modern design proposed at:
https://bugzilla.gnome.org/show_bug.cgi?id=711252

It still uses the sidebar to provide access to multiple
places (keyrings), while backends are moved to the stack switcher on the
headerbar.  "Secure Shell" and "PGP Keys" are folded into "Keys" page.

The sidebar and the key list implementations haven't been touched yet
and could be improved.